+++ examples/options.js
@@ -4,6 +4,7 @@ const options = {
 	url: 'https://github.com/BetaHuhn/metadata-scraper',
 	maxRedirects: 0, // default: 5
 	ua: 'MyApp',
+	lang: 'de-CH',
 	timeout: 1000, // default: 10000
 	forceImageHttps: false, // default: true
 	customRules: {}