+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- lang option which sets the `Accept-Language` header used for the request #12
+
+### Changed
+- update dependencies
+
 ## [v0.1.2] - 2020-12-08
 ### Changed
 - update dependencies